move mpc into contracts directory
