+++ src/chains/near.ts
@@ -3,7 +3,7 @@ import { keyStores, KeyPair, connect, Account } from "near-api-js";
 import { Wallet } from "@near-wallet-selector/core";
 
 export const TGAS = new BN(1000000000000);
-export const NO_DEPOSIT = "0";
+export const NO_DEPOSIT = new BN("0");
 
 export interface NearConfig {
   networkId: string;
