+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from "./chains/ethereum";
-export * from "./mpcContract";
+export * from "./contracts/mpc";
 export * from "./chains/near";
 export * from "./types";
 export * from "./network";
